+++ src/domain/events/queries/EventQueries.ts
@@ -49,6 +49,7 @@ export const eventQuery = gql`
       }
       capacityPerOccurrence
       publishedAt
+      readyForEventGroupPublishing
       eventGroup {
         id
         name